pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "Cython>=3.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deepwater"
version = "0.0.1"
description = "Zero-copy market data substrate with microsecond precision"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{ name = "Dan Gove", email = "dnlgove2@gmail.com" }]
keywords = ["market-data", "time-series", "zero-copy", "shared-memory"]

dependencies = [
    "Cython>=3.0",
    "numpy>=1.26",
    "numba>=0.61.2",
    "orjson>=3.11.3",
    "fastnumbers>=5.1.1",
    "pysimdjson>=7.0.2",
    "websocket-client>=1.8.0",
    "python-crontab>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "prompt_toolkit>=3.0.52",
]

[project.scripts]
deepwater = "deepwater.cli.bootstrap:main"
deepwater-cleanup = "deepwater.ops.cleanup:main"
deepwater-health = "deepwater.ops.health_check:main"
deepwater-repair = "deepwater.ops.repair:main"
deepwater-create-feed = "deepwater.cli.feed_cli:create_main"
deepwater-delete-feed = "deepwater.cli.feed_cli:delete_main"
deepwater-segments = "deepwater.cli.segments_cli:main"
deepwater-datasets = "deepwater.cli.datasets_cli:main"

[project.urls]
Homepage = "https://github.com/DanielGove/Deepwater"
Repository = "https://github.com/DanielGove/Deepwater"
Issues = "https://github.com/DanielGove/Deepwater/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["deepwater", "deepwater.*"]
